refactor: [#37] :truck: Rename question to quiz
